base: error: Do not raise signals unless runinng on GDB
The debug build raises a signal whenever an assert fails, which should be
captured by the debugger making it possible to inspect the cause of the
failure. However, when running the debug build as a standalone process,
raising that signal would cause the application to exit.

Since GDB (seems to) intercept some signals (like SIGTRAP and SIGINT),
setting a flag on these handlers may be (in a hack-y way) used to detect
whether the application in running in the debugger or not. After that,
only raise signals if running on GDB.