base: menu_ctx: Increase contrast and readability between all options modes
Increase the constract between options by adding a new font color, with
really low contrast. If the cursor is on a given line, the active and
inactive fonts are used (which also had their contrast increased). However,
if the cursor isn't on a line with options, it uses the inactive font for
the currently selected option and the "non-selected" font for the others.